comments added in INotificationHandler
